\name{shoes}
\alias{shoes}
\docType{data}
\title{Boys' shoes data}
\description{Data for the boys' shoes example in Box et al. (2005, 81--84).
}
\usage{data(shoes)}
\format{
  A data frame with 10 observations on the following 6 variables.
  \describe{
    \item{\code{boy}}{a numeric vector}
    \item{\code{matA}}{a numeric vector}
    \item{\code{sideA}}{a factor with levels \code{L} \code{R}}
    \item{\code{matB}}{a numeric vector}
    \item{\code{sideB}}{a factor with levels \code{L} \code{R}}
    \item{\code{diff}}{a numeric vector}
  }
}
\details{}
\source{Box, G.E.P, Hunter, J.S. and Hunter, G.W. (2005), \emph{Statistics for Experimenters: Design, Innovation, and Discovery. Second Edition}. Hoboken, NJ: Wiley.}
\references{
}
\examples{
data(shoes)
print(shoes)
}
\keyword{datasets}
