\name{tomatoes}
\alias{tomatoes}
\docType{data}
\title{Tomato yield example}
\description{Data for the Tomato yield example in Box et al. (2005, 78--80)}
\usage{data(tomatoes)}
\format{
  A data frame with 11 observations on the following 4 variables.
  \describe{
    \item{\code{run}}{a numeric vector}
    \item{\code{position}}{a numeric vector}
    \item{\code{fertilizer}}{a factor with levels \code{A} \code{B}}
    \item{\code{pounds}}{a numeric vector}
  }
}
\details{}
\source{Box, G.E.P, Hunter, J.S. and Hunter, G.W. (2005), \emph{Statistics for Experimenters: Design, Innovation, and Discovery. Second Edition}. Hoboken, NJ: Wiley.}
\references{}
\examples{
data(tomatoes)
print(tomatoes)
}
\keyword{datasets}
